requirements.txt
Django==1.8.2
djangorestframework==3.1.3
markdown==2.6.2
django-filter==0.10.0
django-cors-headers==1.1.0